requirements.txt
blinker==1.9.0
click==8.2.1
Flask==3.1.1
Flask-Mail==0.10.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
python-dotenv==1.1.0
Werkzeug==3.1.3